+++ src/components/App/App.js
@@ -1,10 +1,10 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "../../images/logo.svg";
+import "./App.css";
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
+      <header>
         <img src={logo} className="App-logo" alt="logo" />
         <p>
           Edit <code>src/App.js</code> and save to reload.
